requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
httpx>=0.27.0
python-multipart>=0.0.9

# ── PDF parsing (IBM Docling) ─────────────────────────────────────────────────
docling>=2.0.0

# ── LLM scoring (VADER + SentenceTransformer) ────────────────────────────────
vaderSentiment>=3.3.2
sentence-transformers>=3.0.0

# ── Speech-to-text ───────────────────────────────────────────────────────────
openai-whisper>=20231117

# ── Video analysis & gaze tracking ───────────────────────────────────────────
numpy>=1.26.0
scipy>=1.13.0
mediapipe>=0.10.0
opencv-python-headless>=4.9.0
deepface>=0.0.93

# ── Database — Google Drive ───────────────────────────────────────────────────
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0

# ── Database — Supabase ───────────────────────────────────────────────────────
supabase>=2.4.0

# ── Auth / password hashing ───────────────────────────────────────────────────
passlib[bcrypt]>=1.7.4

# ── Environment ───────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
